Refactor signer/core so that EIP712 methods needed by Mesh can compile to WASM
